Handle Quit such that the output file is closed nicely
The previous architecture made it impossible to write out a closing
XML tag in the output file if the GUI parent exits before all
downloads are done (e.g., due to the user clicking the Quit button).
The only way I could see it done is to make the download loop
interruptible, which requires a Thread, and then some additional
architectural changes in the __main__ and control file code to invoke
stop() at the appropriate time.