Fixed handling quit
Quit was not doing the right thing with destroying the frame.  It
needs to call the main thread stop() method and needs to do it in a
subthread to avoid hanging the GUI.  This was ridiculously hard to
solve but in the end did not require many lines of code.